The lnarray size was increased to (600,2) from (400,2) to accomodate crops that didn't mature until the next year and went over 400 days. This occurred in initday.for, leafno.for, output.for and PhenologyMMS.for.
Some comments were added or edited in various files. Some new tinput.dat files were added. Some print statements were commented out.

 The last stage in which leaves would be added was changed from antss (anthesis starts) to epods (early pod stage) in the output for Dry Beans in output.for. The index for the write statement for Soybean in output.for was corrected.

Science changes were made to accomodate the changes made in the Growth Stages for Sorghum, Hay Millet and Proso Millet.

The growth stage of E-SR (Emergence - Single Ridge) in Hay Millet was divided into E - TI (First Tiller) and TI - SR. The addition of GDD's for stage's to reach the next stage were adjusted throughout the phenolhm.for code. The same was done for Proso Millet in phenolpm.for.

Sorghum was redesigned so that the growth stages were more linear. The indices of some of the stages were changed in output.for to match the new sequence.
In phenolsg.for, code was added for the stages First Tiller (tis), Internode Elongation Starts (ies) and Jointing (j). They had previously been included in a span of stages.The addition of GDD's of stages to reach the next stage's GDD amount was revised throughout the code as necessary. Deleted a GoTo statement, adjusted Print statements, added comments and commented out unneeded sections.

In setup.for, the stages to be added together to reach phase one of canopy height were adjusted for Hay Millet, Proso Millet and Sorghum. The stages to be added together to reach phase two of canopy height in Sorghum were adjusted.